+++ admin/content/data_paket.php
@@ -35,6 +35,7 @@
                         $no = 1;
                         $query = mysqli_query($connect, "SELECT * FROM `package`");
                         while ($package = mysqli_fetch_array($query)) {
+                          $id = $package["id"];
                       ?>
                       <tr role="row" class="odd">
                       <td class=""><?php echo $no;$no++; ?></td>
@@ -43,8 +44,36 @@
                       <td class=""> <?php echo $package["duration"]; ?> Days</td>
                       <td class="sorting_1">Rp <?php echo $package["price"]; ?></td>
                       <td class="">
+                            <button type="button" class="btn btn-success" data-toggle="modal" data-target="#modal-default<?php echo $id; ?>">View</button>
+                            <div class="modal fade" id="modal-default<?php echo $id; ?>">
+                            <div class="modal-dialog">
+                              <div class="modal-content">
+                                <div class="modal-header">
+                                  <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                                    <span aria-hidden="true">&times;</span></button>
+                                  <h4 class="modal-title">Detail Package</h4>
+                                </div>
+                                <div class="modal-body" style="padding:20px">
+                                  <div class="row">
+                                    <div class="col-md-6">
+                                      <img src="proses/<?php echo $package["photo"]; ?>" style="width:100%">
+                                    </div>
+                                    <div class="col-md-6">
+                                        <h2 style="margin-top: 0px;"><?php echo $package["name"]; ?></h2>
+                                        <p style="font-size:12px"><?php echo $package["duration"]; ?> Day | <?php echo $package["person"]; ?> Person</p>
+                                        <p><?php echo $package["description"]; ?></p>
+                                        <h4>Rp <?php echo $package["price"]; ?></h4>
+                                      </div>
+                                    </div>
+                                </div>
+                              </div>
+                              <!-- /.modal-content -->
+                            </div>
+                            <!-- /.modal-dialog -->
+                          </div>
+                        </div>
                         <button type="button" class="btn btn-primary" name="button"> Edit</button>
-                        <a href="proses/delete.php?data=package&id=<?php echo $package['id']; ?>&page=data_paket" type="button" class="btn btn-danger" name="button">Delete</a>
+                        <a onclick="return confirm('Are you sure you want to delete this item?');" href="proses/delete.php?data=package&id=<?php echo $package['id']; ?>&page=data_paket" type="button" class="btn btn-danger" name="button">Delete</a>
                       </td>
                     </tr>
                       <?php
